+++ src_c/peripheral.cpp
@@ -328,6 +328,51 @@ simpleble_err_t simpleble_peripheral_unsubscribe(simpleble_peripheral_t handle, 
     return success ? SIMPLEBLE_SUCCESS : SIMPLEBLE_FAILURE;
 }
 
+simpleble_err_t simpleble_peripheral_read_descriptor(simpleble_peripheral_t handle, simpleble_uuid_t service,
+                                                     simpleble_uuid_t characteristic, simpleble_uuid_t descriptor,
+                                                     uint8_t** data, size_t* data_length) {
+    if (handle == nullptr || data == nullptr || data_length == nullptr) {
+        return SIMPLEBLE_FAILURE;
+    }
+
+    // Clear the initial values for safety
+    *data = nullptr;
+    *data_length = 0;
+
+    // Perform the read operation
+    SimpleBLE::Safe::Peripheral* peripheral = (SimpleBLE::Safe::Peripheral*)handle;
+    std::optional<SimpleBLE::ByteArray> read_data = peripheral->read(SimpleBLE::BluetoothUUID(service.value),
+                                                                     SimpleBLE::BluetoothUUID(characteristic.value),
+                                                                     SimpleBLE::BluetoothUUID(descriptor.value));
+
+    // Early return if the operation failed
+    if (!read_data.has_value()) {
+        return SIMPLEBLE_FAILURE;
+    }
+
+    *data_length = read_data.value().size();
+    *data = static_cast<uint8_t*>(malloc(*data_length));
+    memcpy(*data, read_data.value().c_str(), *data_length);
+
+    return SIMPLEBLE_SUCCESS;
+}
+
+simpleble_err_t simpleble_peripheral_write_descriptor(simpleble_peripheral_t handle, simpleble_uuid_t service,
+                                                      simpleble_uuid_t characteristic, simpleble_uuid_t descriptor,
+                                                      const uint8_t* data, size_t data_length) {
+    if (handle == nullptr || data == nullptr) {
+        return SIMPLEBLE_FAILURE;
+    }
+
+    SimpleBLE::Safe::Peripheral* peripheral = (SimpleBLE::Safe::Peripheral*)handle;
+
+    bool success = peripheral->write(
+        SimpleBLE::BluetoothUUID(service.value), SimpleBLE::BluetoothUUID(characteristic.value),
+        SimpleBLE::BluetoothUUID(descriptor.value), SimpleBLE::ByteArray((const char*)data, data_length));
+
+    return success ? SIMPLEBLE_SUCCESS : SIMPLEBLE_FAILURE;
+}
+
 simpleble_err_t simpleble_peripheral_set_callback_on_connected(simpleble_peripheral_t handle,
                                                                void (*callback)(simpleble_peripheral_t, void*),
                                                                void* userdata) {